+++ crates/holochain/src/conductor/conductor.rs
@@ -157,6 +157,9 @@ where
 
     /// Handle to the network actor.
     holochain_p2p: holochain_p2p::HolochainP2pRef,
+
+    /// Database sync level
+    db_sync_level: DbSyncLevel,
 }
 
 impl Conductor {
@@ -473,10 +476,11 @@ where
             Some(env) => Ok(env.clone()),
             None => {
                 let dir = self.root_env_dir.clone();
-                let env = EnvWrite::open(
+                let env = EnvWrite::open_with_sync_level(
                     dir.as_ref(),
                     DbKind::Cache(dna_hash.clone()),
                     self.keystore.clone(),
+                    self.db_sync_level,
                 )?;
                 caches.insert(dna_hash.clone(), env.clone());
                 Ok(env)
@@ -614,16 +618,18 @@ where
             let root_env_dir = root_env_dir.clone();
             let keystore = keystore.clone();
             let conductor_handle = conductor_handle.clone();
+            let db_sync_level = self.db_sync_level;
             async move {
                 use holochain_p2p::actor::HolochainP2pRefToCell;
                 let holochain_p2p_cell = self
                     .holochain_p2p
                     .to_cell(cell_id.dna_hash().clone(), cell_id.agent_pubkey().clone());
 
-                let env = EnvWrite::open(
+                let env = EnvWrite::open_with_sync_level(
                     &root_env_dir,
                     DbKind::Cell(cell_id.clone()),
                     keystore.clone(),
+                    db_sync_level,
                 )
                 .map_err(|err| (cell_id.clone(), err.into()))?;
 
@@ -742,17 +748,18 @@ where
                 let app_id = app_id.clone();
                 let slot_id = app_id.clone();
                 move |mut state| {
-                if let Some(app) = state.installed_apps_mut().get_mut(&app_id) {
-                    let slot = app
-                        .slots()
-                        .get(&slot_id)
-                        .ok_or_else(|| AppError::SlotIdMissing(slot_id.to_owned()))?;
-                    let parent_dna_hash = slot.dna_hash().clone();
-                    Ok((state, parent_dna_hash))
-                } else {
-                    Err(ConductorError::AppNotRunning(app_id.clone()))
+                    if let Some(app) = state.installed_apps_mut().get_mut(&app_id) {
+                        let slot = app
+                            .slots()
+                            .get(&slot_id)
+                            .ok_or_else(|| AppError::SlotIdMissing(slot_id.to_owned()))?;
+                        let parent_dna_hash = slot.dna_hash().clone();
+                        Ok((state, parent_dna_hash))
+                    } else {
+                        Err(ConductorError::AppNotRunning(app_id.clone()))
+                    }
                 }
-            }})
+            })
             .await?;
         let child_dna = dna_store
             .get(&parent_dna_hash)
@@ -1037,6 +1044,7 @@ pub(super) async fn genesis_cells(
     keystore: KeystoreSender,
     cell_ids_with_proofs: Vec<(CellId, Option<MembraneProof>)>,
     conductor_handle: ConductorHandle,
+    db_sync_level: DbSyncLevel,
 ) -> ConductorResult<()> {
     let cells_tasks = cell_ids_with_proofs
         .into_iter()
@@ -1050,10 +1058,11 @@ pub(super) async fn genesis_cells(
                 .await
                 .map_err(Box::new)?;
             tokio::spawn(async move {
-                let env = EnvWrite::open(
+                let env = EnvWrite::open_with_sync_level(
                     &root_env_dir,
                     DbKind::Cell(cell_id_inner.clone()),
                     keystore.clone(),
+                    db_sync_level,
                 )?;
                 Cell::genesis(cell_id_inner, conductor_handle, env, ribosome, proof).await
             })
@@ -1072,7 +1081,8 @@ pub(super) async fn genesis_cells(
     // If there were errors, cleanup and return the errors
     if !errors.is_empty() {
         for cell_id in success {
-            let db = DbWrite::open(&root_env_dir, DbKind::Cell(cell_id))?;
+            let db =
+                DbWrite::open_with_sync_level(&root_env_dir, DbKind::Cell(cell_id), db_sync_level)?;
             db.remove().await?;
         }
 
@@ -1144,6 +1154,7 @@ where
         keystore: KeystoreSender,
         root_env_dir: EnvironmentRootPath,
         holochain_p2p: holochain_p2p::HolochainP2pRef,
+        db_sync_level: DbSyncLevel,
     ) -> ConductorResult<Self> {
         Ok(Self {
             conductor_env: env,
@@ -1158,6 +1169,7 @@ where
             keystore,
             root_env_dir,
             holochain_p2p,
+            db_sync_level,
         })
     }
 
@@ -1326,11 +1338,19 @@ mod builder {
             };
             let env_path = self.config.environment_path.clone();
 
-            let environment =
-                EnvWrite::open(env_path.as_ref(), DbKind::Conductor, keystore.clone())?;
+            let environment = EnvWrite::open_with_sync_level(
+                env_path.as_ref(),
+                DbKind::Conductor,
+                keystore.clone(),
+                self.config.db_sync_level,
+            )?;
 
-            let wasm_environment =
-                EnvWrite::open(env_path.as_ref(), DbKind::Wasm, keystore.clone())?;
+            let wasm_environment = EnvWrite::open_with_sync_level(
+                env_path.as_ref(),
+                DbKind::Wasm,
+                keystore.clone(),
+                self.config.db_sync_level,
+            )?;
 
             #[cfg(any(test, feature = "test_utils"))]
             let state = self.state;
@@ -1361,6 +1381,7 @@ mod builder {
                 keystore,
                 env_path,
                 holochain_p2p,
+                config.db_sync_level,
             )
             .await?;
 
@@ -1377,6 +1398,7 @@ mod builder {
                 conductor: RwLock::new(conductor),
                 keystore,
                 holochain_p2p,
+                db_sync_level: config.db_sync_level,
 
                 #[cfg(any(test, feature = "test_utils"))]
                 skip_publish: std::sync::atomic::AtomicBool::new(false),
@@ -1465,6 +1487,7 @@ mod builder {
                 keystore,
                 self.config.environment_path.clone(),
                 holochain_p2p,
+                self.config.db_sync_level,
             )
             .await?;
 
@@ -1482,6 +1505,7 @@ mod builder {
                 holochain_p2p,
                 p2p_env: envs.p2p(),
                 p2p_metrics_env: envs.p2p_metrics(),
+                db_sync_level: self.config.db_sync_level,
                 #[cfg(any(test, feature = "test_utils"))]
                 skip_publish: std::sync::atomic::AtomicBool::new(false),
             });