+++ crates/kitsune_p2p/kitsune_p2p/src/gossip/sharded_gossip/bloom.rs
@@ -191,23 +191,13 @@ impl ShardedGossipLocal {
     }
 }
 
-// #[cfg(feature = "space_gossip")]
 async fn get_agent_info(
     evt_sender: &EventSender,
     space: &Arc<KitsuneSpace>,
-    _arc_set: Arc<DhtArcSet>,
+    arc_set: Arc<DhtArcSet>,
 ) -> KitsuneResult<Vec<AgentInfoSigned>> {
-    store::all_agent_info(evt_sender, space).await
+    Ok(store::agent_info_within_arc_set(evt_sender, space, arc_set)
+        .await?
+        // Need to collect to know the length for the bloom filter.
+        .collect())
 }
-
-// #[cfg(not(feature = "space_gossip"))]
-// async fn get_agent_info(
-//     evt_sender: &EventSender,
-//     space: &Arc<KitsuneSpace>,
-//     arc_set: Arc<DhtArcSet>,
-// ) -> KitsuneResult<Vec<AgentInfoSigned>> {
-//     Ok(store::agent_info_within_arc_set(evt_sender, space, arc_set)
-//         .await?
-//         // Need to collect to know the length for the bloom filter.
-//         .collect())
-// }