+++ crates/kitsune_p2p/kitsune_p2p/src/spawn/actor/space.rs
@@ -137,6 +137,10 @@ impl SpaceInternalHandler for Space {
         let bootstrap_service = self.config.bootstrap_service.clone();
         let expires_after = self.config.tuning_params.agent_info_expires_after_ms as u64;
         let dynamic_arcs = self.config.tuning_params.gossip_dynamic_arcs;
+        let single_storage_arc_per_space = self
+            .config
+            .tuning_params
+            .gossip_single_storage_arc_per_space;
         let internal_sender = self.i_s.clone();
         Ok(async move {
             let urls = vec![bound_url.into()];
@@ -154,6 +158,7 @@ impl SpaceInternalHandler for Space {
                     mdns_handles: &mut mdns_handles,
                     bootstrap_service: &bootstrap_service,
                     dynamic_arcs,
+                    single_storage_arc_per_space,
                 };
                 peer_data.push(update_single_agent_info(input).await?);
             }
@@ -182,6 +187,10 @@ impl SpaceInternalHandler for Space {
         let bootstrap_service = self.config.bootstrap_service.clone();
         let expires_after = self.config.tuning_params.agent_info_expires_after_ms as u64;
         let dynamic_arcs = self.config.tuning_params.gossip_dynamic_arcs;
+        let single_storage_arc_per_space = self
+            .config
+            .tuning_params
+            .gossip_single_storage_arc_per_space;
         let arc = self.get_agent_arc(&agent);
 
         Ok(async move {
@@ -198,6 +207,7 @@ impl SpaceInternalHandler for Space {
                 mdns_handles: &mut mdns_handles,
                 bootstrap_service: &bootstrap_service,
                 dynamic_arcs,
+                single_storage_arc_per_space,
             };
             let peer_data = vec![update_single_agent_info(input).await?];
             evt_sender
@@ -344,6 +354,7 @@ struct UpdateAgentInfoInput<'borrow> {
     mdns_handles: &'borrow mut HashMap<Vec<u8>, Arc<AtomicBool>>,
     bootstrap_service: &'borrow Option<Url2>,
     dynamic_arcs: bool,
+    single_storage_arc_per_space: bool,
 }
 
 async fn update_arc_length(
@@ -371,9 +382,13 @@ async fn update_single_agent_info(
         mdns_handles,
         bootstrap_service,
         dynamic_arcs,
+        single_storage_arc_per_space,
     } = input;
 
-    if dynamic_arcs {
+    // If there is only a single agent per space don't update the empty arcs.
+    let should_not_update_arc_length = single_storage_arc_per_space && arc.is_empty();
+
+    if dynamic_arcs && !should_not_update_arc_length {
         update_arc_length(evt_sender, space.clone(), &mut arc).await?;
     }
 
@@ -1098,35 +1113,31 @@ impl Space {
         .into())
     }
 
-    // #[cfg(not(feature = "space_gossip"))]
-    // /// Get the existing agent storage arc or create a new one.
-    // fn get_agent_arc(&self, agent: &Arc<KitsuneAgent>) -> DhtArc {
-    //     // TODO: We are simply setting the initial arc to full.
-    //     // In the future we may want to do something more intelligent.
-    //     self.agent_arcs
-    //         .get(agent)
-    //         .cloned()
-    //         .unwrap_or_else(|| DhtArc::full(agent.get_loc()))
-    // }
-
-    // #[cfg(feature = "space_gossip")]
     /// Get the existing agent storage arc or create a new one.
     fn get_agent_arc(&self, agent: &Arc<KitsuneAgent>) -> DhtArc {
-        // TODO: This is an experimental feature that sets the first
-        // agent to join as the full arc and all other later
-        // agents to empty.
-        // It should not be used in production unless you understand
-        // what you are doing.
-        let arc = self.agent_arcs.get(agent).cloned();
-        match arc {
-            Some(arc) => arc,
-            None => {
-                if self.agent_arcs.is_empty() {
-                    DhtArc::full(agent.get_loc())
-                } else {
-                    DhtArc::empty(agent.get_loc())
+        if self
+            .config
+            .tuning_params
+            .gossip_single_storage_arc_per_space
+        {
+            let arc = self.agent_arcs.get(agent).cloned();
+            match arc {
+                Some(arc) => arc,
+                None => {
+                    if self.agent_arcs.is_empty() {
+                        DhtArc::full(agent.get_loc())
+                    } else {
+                        DhtArc::empty(agent.get_loc())
+                    }
                 }
             }
+        } else {
+            // TODO: We are simply setting the initial arc to full.
+            // In the future we may want to do something more intelligent.
+            self.agent_arcs
+                .get(agent)
+                .cloned()
+                .unwrap_or_else(|| DhtArc::full(agent.get_loc()))
         }
     }
 }